chore: moved list component
