+++ src/views/CreateProfile/ChooseProfession.tsx
@@ -3,7 +3,7 @@ import { useQuery, useMutation } from 'react-apollo-hooks'
 import gql from 'graphql-tag'
 import { TaxonomyType } from '../../generated/myskills.d'
 import { RouteComponentProps } from '@reach/router'
-import ExperienceList from '../../components/ExperienceList/ExperienceList'
+import ExperienceList from '../../components/ExperienceList'
 import Flex from '../../components/Flex'
 import Input from '../../components/Input'
 import { H1, H3 } from '../../components/Typography'
